pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pymediaparser"
version = "0.1.0"
description = "媒体流解析与视觉语言模型工具包"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}

# 核心依赖：运行 Pipeline 必需的最小依赖集
dependencies = [
    "numpy>=1.24",
    "pillow>=10.0",
    "av>=12.0",
    "opencv-python>=4.8",
    "requests>=2.28",
]

# 可选依赖：按功能分组，用户按需安装
[project.optional-dependencies]
# VLM 后端依赖
vlm-qwen = [
    # Qwen2-VL / Qwen3-VL / Qwen3.5 本地推理
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.40",
    "qwen-vl-utils>=0.0.8",
]
vlm-qwen3 = [
    # Qwen3-VL / Qwen3.5 需要 transformers>=4.57
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.57.0",
    "qwen-vl-utils>=0.0.8",
]
# 智能采样功能
smart-sampling = [
    "imagehash>=4.3",
    "scikit-image>=0.21",
]
# 开发依赖
dev = [
    "pytest>=8.0",
]
# 便捷安装：常用组合
all = [
    "pymediaparser[vlm-qwen,smart-sampling,dev]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.pip]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"